+++ examples/hono-official-examples/src/examples/custom-event.tsx
@@ -19,7 +19,7 @@ example.get("/", () =>
           id="custom-event-target"
           class="event-output"
           {...state.attrs()}
-          {...ds.on("myevent", ds.expr`${state.$._eventDetails} = evt.detail`)}
+          {...ds.on("myevent", ds.set(state.$._eventDetails, ds.expr`evt.detail`))}
           {...ds.text(ds.expr`${"Last Event Details: "} + ${state.$._eventDetails}`)}
         ></p>
       </div>